+++ app.py
@@ -1,3 +1,9 @@
+__version__ = '22.05.04.1'
+__author__ = 'Oren Brigg & Ramona Renner'
+__author_email__ = 'obrigg@cisco.com / ramrenne@cisco.com'
+__license__ = "Cisco Sample Code License, Version 1.1 - https://developer.cisco.com/site/license/cisco-sample-code-license/"
+
+
 """ Copyright (c) 2021 Cisco and/or its affiliates.
 This software is licensed to you under the terms of the Cisco Sample
 Code License, Version 1.1 (the "License"). You may obtain a copy of the
@@ -15,7 +21,6 @@ or implied.
 from flask import Flask, render_template, request, url_for, redirect, session
 from requests.auth import HTTPBasicAuth
 import backend
-import mab_cleanup
 from time import ctime, sleep, time
 from threading import Thread
 
@@ -24,30 +29,22 @@ app = Flask(__name__)
 app.secret_key = "CiscoISEShared"
 
 # Methods
-
-
 def convert_voucher_list(voucher_list):
     '''
     This function will receive a voucher list with MAC addresses (in "the cisco way" format)
     and expire timestamps (in seconds since the epoch, in UTC) and convert it
     to MAC addresses (in the format xx:xx:xx:xx:xx:xx) and expire dates (as strings in local time).
-    In: {"xxxx.xxxx.xxxx": 1613129301.6337228, "xxxx.xxxx.xxxx": 1613129332.6337228}
-    Out:[{'MACAddress': 'xx:xx:xx:xx:xx:xx', 'ExpDate': 'Fri Feb 12 12:28:21 2021', 'group': 'AAA-Vouchers', 'user': 'Oren'},
-        {'MACAddress': 'xx:xx:xx:xx:xx:xx', 'ExpDate': 'Fri Feb 12 18:28:21 2021', 'group': 'BBB-Vouchers', 'user': 'Ramona'}]
     '''
-    converted_voucher_list = []
-
     for voucher in voucher_list:
 
-        dot_free_mac = voucher['mac'].replace('.', '')
-        mac_address = ':'.join(dot_free_mac[i:i+2] for i in range(0, 12, 2))
-        exp_date = ctime(voucher['duration'])
-        voucher_group = voucher['group']
-        user = voucher['user']
-        converted_voucher_list.append(
-            {"MACAddress": mac_address, "ExpDate": exp_date, "group": voucher_group, "user": user})
+        if voucher['type'] == "host":
+            voucher['mac'] = normalize_mac_format(voucher['mac'])
+            voucher['ExpDate'] = ctime(voucher['duration'])
 
-    return converted_voucher_list
+        elif voucher['type'] == "port":
+            voucher['ExpDate'] = ctime(voucher['duration'])
+
+    return voucher_list
 
 
 def propagate_backend_exception(backend_response):
@@ -64,6 +61,19 @@ def voucher_cleanup_loop():
         backend.voucher_cleanup()
         sleep(10*60)
 
+
+def normalize_mac_format(mac_address):
+    '''
+    This function normalizes mac addresses of different formats 
+    e.g. xxxxxxxxxxxx, xx-xx-xx-xx-xx-xx, xx.xx.xx.xx.xx.xx or xxxx.xxxx.xxxx
+    to xx:xx:xx:xx:xx:xx
+    '''
+    #Remove all :, - and . 
+    mac_address = mac_address.replace('.','').replace(':','').replace('-','')
+    #Add : after every second char
+    normalized_mac_address = ':'.join(a+b for a,b in zip(mac_address[::2], mac_address[1::2]))
+
+    return normalized_mac_address
 
 # Routes
 @app.route('/', methods=['GET'])
@@ -104,17 +114,6 @@ def deviceList():
 
                 return render_template('deviceList.html', device_list=device_list)
 
-            elif request.method == 'POST':
-
-                ip_address = request.form.get("ip_address")
-                relevant_sessions = backend.get_device_auth_sessions(ip_address)
-                propagate_backend_exception(relevant_sessions)
-
-                print(relevant_sessions)
-                print(ip_address)
-
-                return render_template('deviceQuery.html', post_request_done=True, ip_address=ip_address, relevant_sessions=relevant_sessions)
-
         except Exception as e:
             print(e)
             return render_template('deviceList.html', error=True, errorcode=e, reloadlink='/')
@@ -122,6 +121,7 @@ def deviceList():
         print("First you need to login")
         return redirect(url_for('login'))
 
+app.route('/deviceQuery?ip_address=<ip_address>')
 @app.route('/deviceQuery', methods=['GET', 'POST'])
 def deviceQuery():
     '''
@@ -133,20 +133,129 @@ def deviceQuery():
         print(f"Login for {session['username']} extended until: {ctime(session['logged_in'])}")
         try:
             if request.method == 'GET':
-
-                return render_template('deviceQuery.html', post_request_done=False, ip_address='', relevant_sessions={})
+                ip_address = request.args.get("ip_address")
 
             elif request.method == 'POST':
-
                 ip_address = request.form.get("ip_address")
+
+            if ip_address == None:
+                ip_address = ''
+                relevant_sessions={}  
+            else:
                 relevant_sessions = backend.get_device_auth_sessions(ip_address)
                 propagate_backend_exception(relevant_sessions)
 
-                return render_template('deviceQuery.html', post_request_done=True, ip_address=ip_address, relevant_sessions=relevant_sessions)
+            return render_template('deviceQuery.html', ip_address=ip_address, relevant_sessions=relevant_sessions)
 
         except Exception as e:
             print(e)
             return render_template('deviceQuery.html', error=True, errorcode=e)
+    else:
+        print("First you need to login")
+        return redirect(url_for('login'))
+
+
+app.route('/switchView?ip_address=<ip_address>&success=<success>')
+@app.route('/switchView', methods=['GET', 'POST'])
+def switchView():
+    '''
+    This function shows an graphical representation of a switch and its port status. 
+    '''
+    if type(session.get('logged_in')) == int and session.get('logged_in') > int(time()):
+        session['logged_in'] = int(time()) + backend.timeout
+        print(f"Login for {session['username']} extended until: {ctime(session['logged_in'])}")
+        try:    
+            if request.method == 'GET':
+                
+                ip_address = request.args.get("ip_address")
+
+            elif request.method == 'POST':
+
+                ip_address = request.form.get("ip_address")
+            
+            success = request.args.get("success")
+
+            if ip_address == None:
+                ip_address = ''
+                detailed_switch_status={} 
+            else:
+                detailed_switch_status = backend.get_device_ports(ip_address)
+                propagate_backend_exception(detailed_switch_status)
+   
+            return render_template('switchView.html', ip_address=ip_address, success=success, detailed_switch_status=detailed_switch_status)
+
+        except Exception as e:
+            print(e)
+            return render_template('switchView.html', error=True, errorcode=e)
+    else:
+        print("First you need to login")
+        return redirect(url_for('login'))
+
+
+app.route('/portAction?ip_address=<ip_address>&interface=<interface>&action=<action>')
+@app.route('/portAction', methods=['GET', 'POST'])
+def portAction():
+    '''
+    This function handles the port actions: clear port sessions and bypass.
+    '''
+    if type(session.get('logged_in')) == int and session.get('logged_in') > int(time()):
+        session['logged_in'] = int(time()) + backend.timeout
+        print(f"Login for {session['username']} extended until: {ctime(session['logged_in'])}")
+        try: 
+            ip_address = request.args.get("ip_address")
+            interface = request.args.get("interface")
+            action = request.args.get("action")
+
+            if action == 'clear':
+                response = backend.clear_port_auth_sessions(ip_address, interface)
+                propagate_backend_exception(response)
+            elif action == 'bypass':
+                response = backend.add_port_voucher(ip_address, interface, 24, session['username'])
+                propagate_backend_exception(response)
+
+            return redirect('/switchView?ip_address='+ip_address+'&success=True') 
+
+        except Exception as e:
+            print(e)
+            return render_template('switchView.html', error=True, errorcode=e)
+    else:
+        print("First you need to login")
+        return redirect(url_for('login'))
+
+
+app.route('/switchViewDetail?ip_address=<ip_address>&interface=<interface>')
+@app.route('/switchViewDetail', methods=['GET', 'POST'])
+def switchViewDetail():
+    '''
+    This function shows the associated session information for a specific interface and device.
+    '''
+    if type(session.get('logged_in')) == int and session.get('logged_in') > int(time()):
+        session['logged_in'] = int(time()) + backend.timeout
+        print(f"Login for {session['username']} extended until: {ctime(session['logged_in'])}")
+        
+        ip_address = ''
+        interface = ''
+        
+        try:
+            if request.method == 'GET':
+
+                ip_address = request.args.get("ip_address")
+                interface = request.args.get("interface")
+
+                relevant_sessions = backend.get_port_auth_sessions(ip_address, interface)
+                propagate_backend_exception(relevant_sessions)
+
+                return render_template('switchViewDetail.html', ip_address=ip_address, interface=interface, relevant_sessions=relevant_sessions)
+
+        except Exception as e:
+            print(e)
+            if "has no authentication sessions" in repr(e):
+                return render_template('switchViewDetail.html', error=False, ip_address=ip_address, interface=interface, relevant_sessions={})
+            elif "No access sessions on" in repr(e):
+                return render_template('switchViewDetail.html', error=False, no_access_session=True, ip_address=ip_address, interface=interface, relevant_sessions={})
+
+            return render_template('switchViewDetail.html', error=True, errorcode=e)
+
     else:
         print("First you need to login")
         return redirect(url_for('login'))
@@ -172,34 +281,40 @@ def voucher():
                 return render_template('voucher.html', voucher_list=voucher_list, new_voucher=False)
 
             elif request.method == 'POST':
+                new_voucher=False
+                deleted_voucher=False
+                submit_type = request.form.get("voucher_sumbit") #Possible values: add, revoke_host, revoke_port
 
-                submit_type = request.form.get("voucher_sumbit")
-                row_mac_address = request.form.get("voucher_sumbit")
-                form_mac_address = request.form.get("mac_address_field")
-                voucher_duration = request.form.get("voucher_duration")
-                voucher_group = request.form.get("voucher_group")
+                #Add
+                if(submit_type == "add"):
+                    form_mac_address = normalize_mac_format(request.form.get("mac_address_field"))
+                    voucher_duration = request.form.get("voucher_duration")
+                    voucher_group = request.form.get("voucher_group")
 
-                if(submit_type == "Add"):
-
-                    add_response = backend.add_voucher(
+                    response = backend.add_host_voucher(
                         form_mac_address, int(voucher_duration), 
                         voucher_group, session['username'])
-                    propagate_backend_exception(add_response)
-                    voucher_list = convert_voucher_list(
-                        backend.read_voucher_list())
-                    propagate_backend_exception(voucher_list)
+                    new_voucher = True
 
-                    return render_template('voucher.html', voucher_list=voucher_list, new_voucher=True)
+                #Revoke
+                elif(submit_type == "revoke_host"):
+                    host_mac = request.form.get("host_mac")
+                    response = backend.revoke_host_voucher(host_mac, session['username'])
+                    deleted_voucher=True
 
-                else:  # Revoke
+                elif(submit_type == "revoke_port"):
+                    port_switch_ip = request.form.get("port_switch_ip")
+                    port_interface = request.form.get("port_interface")
+                    port_command = request.form.get("port_command")
+                    response = backend.revoke_port_voucher(port_switch_ip, port_interface, port_command, session['username'])
+                    deleted_voucher=True
 
-                    revoke_response = backend.revoke_voucher(row_mac_address, session['username'])
-                    propagate_backend_exception(revoke_response)
-                    voucher_list = convert_voucher_list(
-                        backend.read_voucher_list())
-                    propagate_backend_exception(voucher_list)
+                propagate_backend_exception(response)
+                voucher_list = convert_voucher_list(
+                    backend.read_voucher_list())
+                propagate_backend_exception(voucher_list)
 
-                    return render_template('voucher.html', voucher_list=voucher_list, deleted_voucher=True)
+                return render_template('voucher.html', voucher_list=voucher_list, new_voucher=new_voucher, deleted_voucher=deleted_voucher)
 
         except Exception as e:
             print(e)
@@ -209,6 +324,7 @@ def voucher():
         return redirect(url_for('login'))
 
 
+@app.route('/endpointQuery?mac_address=<mac_address>')
 @app.route('/endpointQuery', methods=['GET', 'POST'])
 def endpointQuery():
     '''
@@ -218,21 +334,32 @@ def endpointQuery():
     if type(session.get('logged_in')) == int and session.get('logged_in') > int(time()):
         session['logged_in'] = int(time()) + backend.timeout
         print(f"Login for {session['username']} extended until: {ctime(session['logged_in'])}")
+        mac_address = ''
+        endpoint_list = ''
         try:
             if request.method == 'GET':
 
-                return render_template('endpointQuery.html', post_request_done=False, endpoint_list={}, mac_address='')
+                mac_address = request.args.get("mac_address")
 
             elif request.method == 'POST':
 
                 mac_address = request.form.get("mac_address")
+
+            if mac_address == None:
+                mac_address = ''
+                endpoint_list={}  
+            else:
+                mac_address=normalize_mac_format(mac_address)
                 endpoint_list = backend.check_ise_auth_status(mac_address)
                 propagate_backend_exception(endpoint_list)
 
-                return render_template('endpointQuery.html', post_request_done=True, endpoint_list=endpoint_list, mac_address=mac_address)
+            return render_template('endpointQuery.html', endpoint_list=endpoint_list, mac_address=mac_address)
 
         except Exception as e:
             print(e)
+            if "No authentication events for MAC Address" in repr(e) and "during the last 24.0 hours." in repr(e):
+                return render_template('endpointQuery.html', error=False, no_events=True, endpoint_list={}, mac_address=mac_address)
+
             return render_template('endpointQuery.html', error=True, errorcode=e)
     else:
         print("First you need to login")